Changes related to STDF import classes
