+++ ATE/Tester/TES/apps/masterApp/resulthandling/stdf_aggregator.py
@@ -3,8 +3,7 @@ import time
 import itertools
 import os
 
-from ATE.data.STDF.MIR import MIR
-from ATE.data.STDF.STDR import STDR
+from ATE.data.STDF import MIR, STDR
 
 # ToDo: Move these methods to somewhere sane, e.g. ATE.data.STDF.utils or similar
 from ATE.Tester.TES.apps.testApp.sequencers.Utils import (generate_FTR, generate_PIR,
