+++ ATE/data/STDF/PIR.py
@@ -1,5 +1,5 @@
 import sys
-from ATE.data.STDF.STDR import STDR
+from ATE.data.STDF import STDR
 
 class PIR(STDR):
     def __init__(self, version=None, endian=None, record = None):
