+++ ATE/data/STDF/__init__.py
@@ -1,6 +1,29 @@
 import os
 
-from ATE.data.STDF.STDR import STDR
-from ATE.data.STDF.STDR import get_STDF_setup_from_file
-from ATE.data.STDF.STDR import ts_to_id
-from ATE.data.STDF.STDR import supported
+from .STDR import *
+from .ATR import *
+from .BPS import *
+from .DTR import *
+from .EPS import *
+from .FAR import *
+from .FTR import *
+from .GDR import *
+from .HBR import *
+from .MIR import *
+from .MPR import *
+from .MRR import *
+from .PCR import *
+from .PGR import *
+from .PIR import *
+from .PLR import *
+from .PMR import *
+from .PRR import *
+from .PTR import *
+from .RDR import *
+from .SBR import *
+from .SDR import *
+from .TSR import *
+from .WCR import *
+from .WIR import *
+from .WRR import *
+from .utils import *
