+++ tests/ATE/data/STDF/test_FAR.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from tests.ATE.data.STDF.STDFRecordTest import STDFRecordTest
-from ATE.data.STDF.FAR import FAR
+from ATE.data.STDF import FAR
 
 #   File Attributes Record
 #   Functuion:
