+++ tests/ATE/data/STDF/test_MIR.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from tests.ATE.data.STDF.STDFRecordTest import STDFRecordTest
-from ATE.data.STDF.MIR import MIR
+from ATE.data.STDF import MIR
 
 #   Master Information Record
 #   Function:
