+++ tests/ATE/data/STDF/test_PGR.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from tests.ATE.data.STDF.STDFRecordTest import STDFRecordTest
-from ATE.data.STDF.PGR import PGR
+from ATE.data.STDF import PGR
 
 #   Pin Group Record
 #   Function:
