+++ tests/ATE/data/STDF/test_PMR.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from tests.ATE.data.STDF.STDFRecordTest import STDFRecordTest
-from ATE.data.STDF.PMR import PMR
+from ATE.data.STDF import PMR
 
 #   Pin Map Record
 #   Function:
