+++ tests/ATE/data/STDF/test_TSR.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from tests.ATE.data.STDF.STDFRecordTest import STDFRecordTest
-from ATE.data.STDF.TSR import TSR
+from ATE.data.STDF import TSR
 
 #   Test Synopsis Record
 #   Function:
