+++ tests/ATE/data/STDF/test_WRR.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from tests.ATE.data.STDF.STDFRecordTest import STDFRecordTest
-from ATE.data.STDF.WRR import WRR
+from ATE.data.STDF import WRR
 
 #   Wafer Results Record
 #   Function:
